backend/requirements.txt
asgiref==3.8.1
boto3==1.37.9
botocore==1.37.9
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
jmespath==1.0.1
pillow==11.1.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
